refactor: move `createWaku` to own module
